押下時の処理を外から入れるように変更
