+++ src/components/blocks/SearchBox.jsx
@@ -5,22 +5,8 @@ import SearchIcon from '@mui/icons-material/Search'
 import SearchIconWrapper from '../atoms/SearchIconWrapper'
 import StyledInputBase from '../atoms/StyledInputBase'
 
-import { useNavigate } from 'react-router-dom'
-
 export default function SearchBox (props) {
-  const navigate = useNavigate()
-
-  const keyPressed = async (event) => {
-    if (event.key !== 'Enter') return
-    event.preventDefault()
-    const searchKeyword = (event.target.value || '').replace(/^\s+|\s+$/g, '')
-    if (searchKeyword === '') return
-    navigate('/searchresult', {
-      state: {
-        searchKeyword: searchKeyword
-      }
-    })
-  }
+  const { placeholder, keyPress } = props
 
   return (
     <Search>
@@ -28,14 +14,15 @@ export default function SearchBox (props) {
         <SearchIcon />
       </SearchIconWrapper>
       <StyledInputBase
-        placeholder={props.placeholder || 'Search...'}
+        placeholder={placeholder || 'Search...'}
         inputProps={{ 'aria-label': 'search' }}
-        onKeyPress={keyPressed}
+        onKeyPress={keyPress}
       />
     </Search>
   )
 }
 
 SearchBox.propTypes = {
-  placeholder: PropTypes.string
+  placeholder: PropTypes.string,
+  keyPress: PropTypes.func
 }
